fix: Revert "js: Add url to wasm file as argument to buttplugInit"
This fixes #88 by reverting commit fe66645f19e43682b55ebfd51a2979036803b4d8

That change was supposed to make it possible to override the location of the
Buttplug JS files, but instead it prevents a webpack path rewriting feature
from working so that without a path, the library is no longer able to find
the files to import.

Worse, due to a mix of JS and WASM and import statements between them, it's
not actually possible to specify a path to a the Buttplug JS files and have
a browser load them: doing so just results in MIME type checking errors.